Clean some bits of this move
