+++ src/lexedata/importer/cognates.py
@@ -5,7 +5,7 @@ import openpyxl
 
 from lexedata import cli
 from lexedata.types import Language, RowObject, CogSet
-import lexedata.importer.cellparser as cell_parsers
+import lexedata.util.excel as cell_parsers
 from lexedata.importer.excel_matrix import ExcelCognateParser
 from lexedata.util.excel import clean_cell_value, get_cell_comment
