+++ src/lexedata/importer/excel_matrix.py
@@ -27,7 +27,7 @@ from lexedata.util.excel import (
     clean_cell_value,
     get_cell_comment,
 )
-import lexedata.importer.cellparser as cell_parsers
+import lexedata.util.excel as cell_parsers
 from lexedata.edit.add_status_column import add_status_column_to_table
 import lexedata.cli as cli
